@@ -10,9 +10,8 @@
             </v-card-title>
             <v-card-text>
                 <v-row justify="center" align="center">
-                    <v-flex>
+                    <v-col>
                         <v-col>
-                            <import-json-motor ref="importForm" @importStart="loading = true"></import-json-motor>
                             <v-alert type="error" v-model="showError" dismissible outlined>
                                 {{ errorMessage }}
                             </v-alert>
@@ -31,16 +30,16 @@
                                 </template>
                                 <template v-slot:item.actions="{ item }">
                                     <v-tooltip bottom>
-                                        <template v-slot:activator="{ on }">
-                                            <v-btn icon v-on="on" @click="loadMotor(item)" text>
+                                        <template v-slot:activator="{ props }">
+                                            <v-btn icon v-bind="props" @click="loadMotor(item)" text>
                                                 <v-icon color="green">mdi-play</v-icon>
                                             </v-btn>
                                         </template>
                                         <span>Run computation</span>
                                     </v-tooltip>
                                     <v-tooltip bottom>
-                                        <template v-slot:activator="{ on }">
-                                            <v-btn icon v-on="on" @click="confirmDelete(item)" text>
+                                        <template v-slot:activator="{ props }">
+                                            <v-btn icon v-bind="props" @click="confirmDelete(item)" text>
                                                 <v-icon color="red">mdi-delete</v-icon>
                                             </v-btn>
                                         </template>
@@ -49,7 +48,7 @@
                                 </template>
                             </v-data-table>
                         </v-col>
-                    </v-flex>
+                    </v-col>
                 </v-row>
             </v-card-text>
             <v-divider></v-divider>
@@ -96,13 +95,11 @@
 <script>
 
 import { mapActions, mapGetters } from 'vuex'
-import ImportJsonMotor from '@/components/motor/ImportJsonMotor'
 import { shortLabel } from '@/modules/utils'
 import { isCustomPropellant } from '@/modules/customPropellant'
 
 export default {
     name: 'MotorSelect',
-    components: { ImportJsonMotor },
     data() {
         return {
             headers: [
